use the correct type for got_bytes. fixes bytestream weirdness taaz was seeing.
Original commit message from CVS:
use the correct type for got_bytes. fixes bytestream weirdness taaz was seeing.

+++ ext/a52dec/gsta52dec.c
@@ -379,7 +379,7 @@ gst_a52dec_loop (GstElement *element)
   int i, length, flags, sample_rate, bit_rate;
   int stream_channels;
   GstBuffer *buf;
-  guint8 got_bytes;
+  guint32 got_bytes;
 
   a52dec = GST_A52DEC (element);
